refactor: update subscription hook import paths
- Changed the import path for the `useSubscription` hook from the `tanstack-query-hooks` directory to the new `hooks` directory for better organization.
- Added the `useSubscription` hook implementation to the `hooks` directory, encapsulating the subscription query logic for improved reusability across the application.
