+++ frontend/src/hooks/use-subscription.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import { getSubscription } from "../api/billing";
+import { getSubscription } from "../lib/api/billing";
 import { useMetaInfo } from "@/components/context/metainfo";
 
 export const useSubscription = () => {
